# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API for bookstore.",
    "title": "Bookstore API GO",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "API Support",
      "url": "http://www.swagger.io/support",
      "email": "support@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0"
  },
  "host": "localhost:9000",
  "basePath": "/api/v1",
  "paths": {
    "/": {
      "get": {
        "description": "do ping",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["example"],
        "summary": "ping example",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/admin/login": {
      "post": {
        "description": "Authenticates an admin using username and password, returns a JWT token if successful",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["admin"],
        "summary": "Authenticate an admin",
        "parameters": [
          {
            "description": "Admin login object",
            "name": "admin",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.LoginAdmin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JWT Token",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/admin/register": {
      "post": {
        "description": "Registers a new admin with the given username and password",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["admin"],
        "summary": "Register a new admin",
        "parameters": [
          {
            "description": "Admin registration object",
            "name": "admin",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.RegisterAdmin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully registered",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/admin/users": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Get all users",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["admin"],
        "summary": "Get all users",
        "responses": {
          "200": {
            "description": "Successfully retrieved all users",
            "schema": {
              "$ref": "#/definitions/models.User"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/admin/users/{username}": {
      "put": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Update user by username",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["admin"],
        "summary": "Update user by username",
        "parameters": [
          {
            "type": "string",
            "description": "Username",
            "name": "username",
            "in": "path",
            "required": true
          },
          {
            "description": "User object",
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.RegisterUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully updated user",
            "schema": {
              "$ref": "#/definitions/models.User"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "string"
            }
          },
          "404": {
            "description": "User not found",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/books": {
      "get": {
        "description": "Get a list of all books with optional pagination",
        "produces": ["application/json"],
        "tags": ["books"],
        "summary": "Get all books with pagination",
        "parameters": [
          {
            "type": "integer",
            "default": 0,
            "description": "Offset for pagination",
            "name": "offset",
            "in": "query"
          },
          {
            "type": "integer",
            "default": 10,
            "description": "Limit for pagination",
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved list of books",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/models.Book"
              }
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "JwtAuth": []
          }
        ],
        "description": "Create a new book with the given input data",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["books"],
        "summary": "Create a new book",
        "parameters": [
          {
            "description": "Create book object",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.CreateBook"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successfully created book",
            "schema": {
              "$ref": "#/definitions/models.Book"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/books/{id}": {
      "get": {
        "description": "Get details of a book by its ID",
        "produces": ["application/json"],
        "tags": ["books"],
        "summary": "Find a book by ID",
        "parameters": [
          {
            "type": "string",
            "description": "Book ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved book",
            "schema": {
              "$ref": "#/definitions/models.Book"
            }
          },
          "404": {
            "description": "Book not found",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "put": {
        "description": "Update the book details for the given ID",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["books"],
        "summary": "Update a book by ID",
        "parameters": [
          {
            "type": "string",
            "description": "Book ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Update book object",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.UpdateBook"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully updated book",
            "schema": {
              "$ref": "#/definitions/models.Book"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "string"
            }
          },
          "404": {
            "description": "book not found",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "delete": {
        "description": "Delete the book with the given ID",
        "produces": ["application/json"],
        "tags": ["books"],
        "summary": "Delete a book by ID",
        "parameters": [
          {
            "type": "string",
            "description": "Book ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully deleted book",
            "schema": {
              "type": "string"
            }
          },
          "404": {
            "description": "book not found",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "description": "Authenticates a user using username and password, returns a JWT token if successful",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["user"],
        "summary": "Authenticate a user",
        "parameters": [
          {
            "description": "User login object",
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.LoginUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JWT Token",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/user/profile": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Get the profile of the currently logged in user",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["user"],
        "summary": "Get user profile",
        "responses": {
          "200": {
            "description": "Successfully retrieved user profile",
            "schema": {
              "$ref": "#/definitions/models.User"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "put": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Update the profile of the currently logged in user",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["user"],
        "summary": "Update user profile",
        "parameters": [
          {
            "description": "User object",
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.RegisterUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully updated user profile",
            "schema": {
              "$ref": "#/definitions/models.User"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/user/register": {
      "post": {
        "description": "Registers a new user with the given username and password",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["user"],
        "summary": "Register a new user",
        "parameters": [
          {
            "description": "User registration object",
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.RegisterUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully registered",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "models.Book": {
      "type": "object",
      "properties": {
        "author": {
          "type": "string"
        },
        "cover": {
          "type": "string"
        },
        "created_at": {
          "type": "string"
        },
        "genre": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "price": {
          "type": "integer"
        },
        "quantity": {
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "updated_at": {
          "type": "string"
        },
        "year": {
          "type": "integer"
        }
      }
    },
    "models.CreateBook": {
      "type": "object",
      "required": [
        "author",
        "cover",
        "genre",
        "price",
        "quantity",
        "title",
        "year"
      ],
      "properties": {
        "author": {
          "type": "string"
        },
        "cover": {
          "type": "string"
        },
        "genre": {
          "type": "string"
        },
        "price": {
          "type": "integer"
        },
        "quantity": {
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "year": {
          "type": "integer"
        }
      }
    },
    "models.LoginAdmin": {
      "type": "object",
      "required": ["password", "username"],
      "properties": {
        "password": {
          "type": "string",
          "example": "admin123!"
        },
        "username": {
          "type": "string",
          "example": "admin"
        }
      }
    },
    "models.LoginUser": {
      "type": "object",
      "required": ["password", "username"],
      "properties": {
        "password": {
          "type": "string",
          "example": "Test123!"
        },
        "username": {
          "type": "string",
          "example": "test"
        }
      }
    },
    "models.RegisterAdmin": {
      "type": "object",
      "required": ["email", "password", "username"],
      "properties": {
        "email": {
          "type": "string",
          "example": "admin@test.com"
        },
        "password": {
          "type": "string",
          "example": "admin123!"
        },
        "username": {
          "type": "string",
          "example": "admin"
        }
      }
    },
    "models.RegisterUser": {
      "type": "object",
      "required": ["email", "password", "username"],
      "properties": {
        "email": {
          "type": "string",
          "example": "test@test.com"
        },
        "password": {
          "type": "string",
          "example": "Test123!"
        },
        "username": {
          "type": "string",
          "example": "test"
        }
      }
    },
    "models.UpdateBook": {
      "type": "object",
      "properties": {
        "author": {
          "type": "string"
        },
        "cover": {
          "type": "string"
        },
        "genre": {
          "type": "string"
        },
        "price": {
          "type": "integer"
        },
        "quantity": {
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "year": {
          "type": "integer"
        }
      }
    },
    "models.User": {
      "type": "object",
      "properties": {
        "created_at": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "password": {
          "type": "string"
        },
        "updated_at": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "externalDocs": {
    "description": "OpenAPI",
    "url": "https://swagger.io/resources/open-api/"
  }
}